+++ src/storage-preview/azext_storage_preview/_params.py
@@ -61,7 +61,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.argument('if_match')
         c.argument('if_none_match')
 
-    for item in ['delete', 'show', 'update', 'show-connection-string', 'keys', 'network-rule', 'failover']:
+    for item in ['delete', 'show', 'update', 'keys', 'network-rule', 'failover']:
         with self.argument_context('storage account {}'.format(item)) as c:
             c.argument('account_name', acct_name_type, options_list=['--name', '-n'])
             c.argument('resource_group_name', required=False, validator=process_resource_group)
@@ -118,13 +118,6 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
     with self.argument_context('storage account show') as c:
         c.argument('expand', arg_type=get_enum_type(
             self.get_models('StorageAccountExpand', resource_type=CUSTOM_MGMT_STORAGE)))
-
-    with self.argument_context('storage account show-connection-string') as c:
-        c.argument('protocol', help='The default endpoint protocol.', arg_type=get_enum_type(['http', 'https']))
-        c.argument('key_name', options_list=['--key'], help='The key to use.',
-                   arg_type=get_enum_type(list(storage_account_key_options.keys())))
-        for item in ['blob', 'file', 'queue', 'table']:
-            c.argument('{}_endpoint'.format(item), help='Custom endpoint for {}s.'.format(item))
 
     with self.argument_context('storage account keys renew') as c:
         c.argument('key_name', options_list=['--key'], help='The key to regenerate.', validator=validate_key,
@@ -198,10 +191,15 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
                    help='Recursively delete blobs.')
         c.ignore('target')
 
-    # with self.argument_context('storage azcopy blob sync') as c:
-    #     c.argument('destination', options_list=['--destination', '-d'],
-    #                validator=validate_azcopy_container_destination_url)
-    #     c.argument('source', options_list=['--source', '-s'])
+    with self.argument_context('storage azcopy blob sync') as c:
+        c.extra('destination_container', options_list=['--container', '-c'], required=True,
+                help='The sync destination container.')
+        c.extra('destination_path', options_list=['--destination', '-d'],
+                validator=validate_azcopy_upload_destination_url,
+                help='The sync destination path.')
+        c.argument('source', options_list=['--source', '-s'],
+                   help='The source file path to sync from.')
+        c.ignore('destination')
 
     with self.argument_context('storage azcopy run-command') as c:
         c.positional('command_args', help='Command to run using azcopy. Please start commands with "azcopy ".')