+++ src/storage-preview/azext_storage_preview/tests/latest/test_storage_account_scenarios.py
@@ -67,11 +67,6 @@ class StorageAccountTests(StorageScenarioMixin, ScenarioTest):
             JMESPathCheck('kind', 'Storage')
         ])
 
-        self.cmd('storage account show-connection-string -g {} -n {} --protocol http'.format(
-            resource_group, name), checks=[
-                JMESPathCheck("contains(connectionString, 'https')", False),
-                JMESPathCheck("contains(connectionString, '{}')".format(name), True)])
-
         self.cmd('storage account update -g {} -n {} --tags foo=bar cat'
                  .format(resource_group, name),
                  checks=JMESPathCheck('tags', {'cat': '', 'foo': 'bar'}))
@@ -102,25 +97,6 @@ class StorageAccountTests(StorageScenarioMixin, ScenarioTest):
             JMESPathCheck('nameAvailable', False),
             JMESPathCheck('reason', 'AlreadyExists')
         ])
-
-    @ResourceGroupPreparer(parameter_name_for_location='location')
-    def test_create_storage_account_file_aad(self, resource_group, location):
-        self.kwargs.update({
-            'name': self.create_random_name(prefix='cli', length=24),
-            'loc': location
-        })
-
-        self.cmd('storage account create -n {name} -g {rg} -l {loc} --kind StorageV2 --file-aad',
-                 checks=[JMESPathCheck('kind', 'StorageV2'),
-                         JMESPathCheck('enableAzureFilesAadIntegration', True)])
-
-        self.cmd('storage account check-name --name {name}', checks=[
-            JMESPathCheck('nameAvailable', False),
-            JMESPathCheck('reason', 'AlreadyExists')
-        ])
-
-        self.cmd('storage account update -n {name} -g {rg} --file-aad false',
-                 checks=[JMESPathCheck('enableAzureFilesAadIntegration', False)])
 
     @ResourceGroupPreparer(parameter_name_for_location='location', location='northeurope')
     def test_create_storage_account_premium_sku(self, resource_group, location):