+++ src/storage-preview/HISTORY.rst
@@ -3,6 +3,26 @@
  Release History
 ===============
 
+0.2.8 (2019-7-5)
+++++++++++++++++
+* Remove file-add command argument
+
+0.2.7 (2019-6-14)
++++++++++++++++++
+* Remove min_profile
+
+0.2.6 (2019-5-28)
++++++++++++++++++
+* Update the max CLI core version as 2.0.66
+
+0.2.5 (2019-5-1)
+++++++++++++++++
+* Release management policy
+
+0.2.4 (2019-4-11)
++++++++++++++++++
+* Release azcopy commands: `az storage azcopy blob sync`
+
 0.2.3 (2019-3-21)
 +++++++++++++++++
 * Release initial azcopy commands: `az storage azcopy blob delete/upload/download`